--- lib/commands/console_delegation.rb
@@ -1,11 +1,0 @@
-require 'commands/commander'
-
-module Commands
-  module ConsoleDelegation
-    def commander
-      @commander ||= Commander.new
-    end
-
-    delegate :rake, :test, :generate, :destroy, :update, to: :commander
-  end
-end
