--- lib/commands/environment.rb
@@ -1,14 +1,0 @@
-module Commands
-  module Environment
-    extend self
-    
-    def fork
-      Kernel.fork do
-        yield
-        Kernel.exit
-      end
-
-      Process.waitall
-    end
-  end
-end
