+++ lib/rails/console/environment.rb
@@ -1,0 +1,16 @@
+module Rails
+  module ConsoleMethods
+    module Environment
+      extend self
+
+      def fork
+        Kernel.fork do
+          yield
+          Kernel.exit
+        end
+
+        Process.waitall
+      end
+    end
+  end
+end
